almost fixed dateTime getter for events, problem is what does a 'null' value input to the table look like? can i get null not 'null' back ever with Ala? this involves vlister and getEventDate
was going to connect genres -> event search artists, but finding that many on spotify just don't have genres listed (I believe, dbl-check my reporting about 'unresolved artists') so I need a way around that problem.

search_artists is having a payload size issue: in a 1-month time frame of events for a metro @ 236 count and 362 (unique? check that) artists, it takes about 36 seconds to resolve (reqs per sec = 11 is max here) which is a problem on its own, obviously won't scale

displaying the events with a payload of that size is difficult, but appears as tho pagination would help